refact: extract print col and row logic to utils
