+++ fakesnake/utils/__init__.py
@@ -1,1 +1,1 @@
-from .environment import *
+from .utils import *
